fix warnings

+++ src/distances.cpp
@@ -4,13 +4,13 @@
 
 double euclidean_distance(std::vector<double> point_a, std::vector<double> point_b)
 {
-    int size = point_a.size();
+    unsigned size = (unsigned)point_a.size();
     if (size != point_b.size())
         throw std::runtime_error("Points with different number of dimensions...");
 
     double summation = 0.0;
 
-    for (int i = 0; i < size; i++)
+    for (unsigned i = 0; i < size; i++)
     {
         summation += pow((point_a[i] - point_b[i]), 2.0);
     }
@@ -20,13 +20,13 @@ double euclidean_distance(std::vector<double> point_a, std::vector<double> point
 
 double manhattan_distance(std::vector<double> point_a, std::vector<double> point_b)
 {
-    int size = point_a.size();
+    unsigned size = (unsigned)point_a.size();
     if (size != point_b.size())
         throw std::runtime_error("Points with different number of dimensions...");
 
     double summation = 0.0;
 
-    for (int i = 0; i < size; i++)
+    for (unsigned i = 0; i < size; i++)
     {
         summation += std::abs(point_a[i] - point_b[i]);
     }
